Added Concerns to Attribute Controller to avoid code conflict issues
